April 26, 2021 Bot API 5.2 ComponentVersion 1.3.4
Support for Payments 2.0, see manual for more details about the Bot Payments API.
Added the type TtdInputInvoiceMessageContent to support sending invoices as inline query results.
Allowed sending invoices to group, supergroup and channel chats.
Added the fields MaxTipAmount and SuggestedTipAmounts to the method sendInvoice to allow adding optional tips to the payment.
The parameter start_parameter of the method sendInvoice became optional. If the parameter isn't specified, the invoice can be paid directly from forwarded messages.
Added the field ChatType to the class TtdInlineQuery, containing the type of the chat, from which the inline request was sent.
Added the type TtdVoiceChatScheduled and the field VoiceChatScheduled to the class TtdMessage.
Fixed an error in sendChatAction documentation to correctly mention “record_voice” and “upload_voice” instead of “record_audio” and “upload_audio” for related to voice note actions. Old action names will still work for backward compatibility.