Component Update for Bot API 6.5 - February 3, 2023
Added requests for users and chats and support for granular media permissions.
Added the class TtdKeyboardButtonRequestUser and the field RequestUser to the class TtdKeyboardButton.
Added the class KeyboardButtonRequestChat and the field request_chat to the class KeyboardButton.
Added the classes TtdUserShared, TtdChatShared and the fields UserShared, and ChatShared to the class TtdMessage.
Replaced the fields CanSendMediaMessages in the classes TtdChatMemberRestricted and TtdChatPermissions with separate fields CanSendAudios, CanSendDocuments, CanSendPhotos, CanSendVideos, CanSendVideoNotes, and CanSendVoiceNotes for different media types.
Added method setChatPermissions.
Added the parameter UseIndependentChatPermissions to the methods restrictChatMember and setChatPermissions.
Added the field user_chat_id to the class TtdChatJoinRequest.