Bot Api Update 6.6 - March 9, 2023
March 9, 2023
Bot API 6.6

Added the ability to set different bot descriptions for different user languages using the method setMyDescription.
Added the ability to get the current bot description in the given language as the class BotDescription using the method getMyDescription.
Added the ability to set different bot short descriptions for different user languages using the method setMyShortDescription.
Added the ability to get the current bot short description in the given language as the class BotShortDescription using the method getMyShortDescription.
Added the parameter emoji to the method sendSticker to specify an emoji for just uploaded stickers.
Added support for the creation of custom emoji sticker sets in createNewStickerSet.
Added the parameter needs_repainting to the method createNewStickerSet to automatically change the color of emoji based on context (e.g., use text color in messages, accent color in statuses, etc.).
Added the field needs_repainting to the class Sticker.
Replaced the parameters png_sticker, tgs_sticker, webm_sticker, emojis and mask_postion in the method addStickerToSet with the parameter sticker of the type InputSticker.
Added support for the creation of sticker sets with multiple initial stickers in createNewStickerSet by replacing the parameters png_sticker, tgs_sticker, webm_sticker, emojis and mask_postion with the parameters stickers and sticker_format.
Added support for .WEBP files in createNewStickerSet and addStickerToSet.
Added support for .WEBP, .TGS, and .WEBM files in uploadStickerFile by replacing the parameter png_sticker in the method uploadStickerFile with the parameters sticker and sticker_format.
Added the ability to specify search keywords for stickers added to sticker sets.
Added the method setCustomEmojiStickerSetThumbnail for editing the thumbnail of custom emoji sticker sets created by the bot.
Added the method setStickerSetTitle for editing the title of sticker sets created by the bot.
Added the method deleteStickerSet for complete deletion of a given sticker set that was created by the bot.
Added the method setStickerEmojiList for changing the list of emoji associated with a sticker.
Added the method setStickerKeywords for changing the search keywords assigned to a sticker.
Added the method setStickerMaskPosition for changing the mask position of a mask sticker.
Renamed the field thumb in the classes Animation, Audio, Document, Sticker, Video, VideoNote, InputMediaAnimation, InputMediaAudio, InputMediaDocument, InputMediaVideo, StickerSet to thumbnail.
Renamed the parameter thumb in the methods sendAnimation, sendAudio, sendDocument, sendVideo, sendVideoNote to thumbnail.
Renamed the method setStickerSetThumb to setStickerSetThumbnail and its parameter thumb to thumbnail.
Renamed the fields thumb_url, thumb_width, and thumb_height in the classes InlineQueryResultArticle, InlineQueryResultContact, InlineQueryResultDocument, InlineQueryResultLocation, and InlineQueryResultVenue to thumbnail_url, thumbnail_width, and thumbnail_height respectively.
Renamed the field thumb_url in the classes InlineQueryResultPhoto and InlineQueryResultVideo to thumbnail_url.
Renamed the fields thumb_url and thumb_mime_type in the classes InlineQueryResultGif, and InlineQueryResultMpeg4Gif to thumbnail_url and thumbnail_mime_type respectively.