December 29, 2023 Bot API 7.0 Update in Component Version 1.4.7
December 29, 2023
Bot API 7.0

Reactions

Added the classes ReactionTypeEmoji and ReactionTypeCustomEmoji representing different types of reaction.
Added updates about a reaction change on a message with non-anonymous reactions, represented by the class MessageReactionUpdated and the field message_reaction in the class Update. The bot must explicitly allow the update to receive it.
Added updates about reaction changes on a message with anonymous reactions, represented by the class MessageReactionCountUpdated and the field message_reaction_count in the class Update. The bot must explicitly allow the update to receive it.
Added the method setMessageReaction that allows bots to react to messages.
Added the field available_reactions to the class Chat.
Replies 2.0

Added the ability to reply to messages in other chats or forum topics.
Added the class ExternalReplyInfo and the field external_reply of type ExternalReplyInfo to the class Message, containing information about a message that is replied to by the current message, but can be from another chat or forum topic.
Added the ability to quote a part of the replied message.
Added the class TextQuote and the field quote of type TextQuote to the class Message, which contains the part of the replied message text or caption that is quoted in the current message.
Added the class ReplyParameters and replaced parameters reply_to_message_id and allow_sending_without_reply in the methods copyMessage, sendMessage, sendPhoto, sendVideo, sendAnimation, sendAudio, sendDocument, sendSticker, sendVideoNote, sendVoice, sendLocation, sendVenue, sendContact, sendPoll, sendDice, sendInvoice, sendGame, and sendMediaGroup with the field reply_parameters of type ReplyParameters.
Link Preview Customization

Allowed to explicitly specify the URL that will be used for link preview generation in outgoing text messages.
Allowed to position link previews above the message text.
Allowed to choose media size in link previews.
Added the class LinkPreviewOptions and replaced the parameter disable_web_page_preview with link_preview_options in the methods sendMessage and editMessageText.
Replaced the field disable_web_page_preview with link_preview_options in the class InputTextMessageContent.
Added the field link_preview_options to the class Message with information about the link preview options used to send the message.
Block Quotation

Added support for “blockquote” entities in received messages.
Added support for “blockquote” entity parsing in “MarkdownV2” and “HTML” parse modes.
Allowed to explicitly specify “blockquote” entities in formatted texts.
Multiple Message Actions

Added the method deleteMessages to allow the deletion of multiple messages in a single request.
Added the method forwardMessages for forwarding of multiple messages in a single request.
Added the method copyMessages for copying of multiple messages in a single request.
Request for multiple users

Renamed the class KeyboardButtonRequestUser to KeyboardButtonRequestUsers and added the field max_quantity to it.
Renamed the field request_user in the class KeyboardButton to request_users. The old name will still work for backward compatibility.
Added the class UsersShared.
Replaced the field user_shared in the class Message with the field users_shared.
Chat Boost

Added updates about chat boost changes, represented by the classes ChatBoostUpdated and ChatBoostRemoved and the fields chat_boost and removed_chat_boost in the class Update. The bot must be an administrator in the chat to receive these updates.
Added the classes ChatBoostSourcePremium, ChatBoostSourceGiftCode and ChatBoostSourceGiveaway, representing different sources of a chat boost.
Added the method getUserChatBoosts for obtaining the list of all active boosts a user has contributed to a chat.
Giveaway

Added the class Giveaway and the field giveaway to the class Message for messages about scheduled giveaways.
Added the class GiveawayCreated and the field giveaway_created to the class Message for service messages about the creation of a scheduled giveaway.
Added the class GiveawayWinners and the field giveaway_winners to the class Message for messages about the completion of a giveaway with public winners.
Added the class GiveawayCompleted and the field giveaway_completed to the class Message for service messages about the completion of a giveaway without public winners.
Web App Changes

Added the field SettingsButton to the class WebApp.
Added the fields header_bg_color, accent_text_color, section_bg_color, section_header_text_color, subtitle_text_color, destructive_text_color to the class ThemeParams.
Web Apps no longer close when the method WebApp.openTelegramLink is called.
Other Changes

Added support for the fields emoji_status_custom_emoji_id and emoji_status_expiration_date in the class Chat for non-private chats.
Added the fields accent_color_id, background_custom_emoji_id, profile_accent_color_id, and profile_background_custom_emoji_id to the class Chat.
Added the field has_visible_history to the class Chat.
Added the class MessageOrigin and replaced the fields forward_from, forward_from_chat, forward_from_message_id, forward_signature, forward_sender_name, and forward_date with the field forward_origin of type MessageOrigin in the class Message.
Improved documentation for the field message of the class callbackQuery and the field pinned_message of the class Message by adding the classes MaybeInaccessibleMessage and InaccessibleMessage.